Add Linux clang build
In the shell script, the CMake CXX flags are updated to use
libc++ (for linking) when compiling with clang++. This leaves an
issue of duplicate flag configuration, one in the CMakeLists and
the other on the bash shell for Travis. Future maintenance or
refacot should address this issue if possible.